LayoutTests/fast/events/constructors/wheel-event-constructor-expected.txt

This tests the constructor for the WheelEvent DOM class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new WheelEvent('eventType').bubbles is false
PASS new WheelEvent('eventType').cancelable is false
PASS new WheelEvent('eventType').view is null
PASS new WheelEvent('eventType').detail is 0
PASS new WheelEvent('eventType').screenX is 0
PASS new WheelEvent('eventType').screenY is 0
PASS new WheelEvent('eventType').clientX is 0
PASS new WheelEvent('eventType').clientY is 0
PASS new WheelEvent('eventType').ctrlKey is false
PASS new WheelEvent('eventType').shiftKey is false
PASS new WheelEvent('eventType').altKey is false
PASS new WheelEvent('eventType').metaKey is false
PASS new WheelEvent('eventType').button is 0
PASS new WheelEvent('eventType').relatedTarget is null
PASS new WheelEvent('eventType').wheelDeltaX is 0
PASS new WheelEvent('eventType').wheelDeltaY is 0
PASS new WheelEvent('eventType').deltaMode is WheelEvent.DOM_DELTA_PIXEL
PASS new WheelEvent('eventType', { bubbles: false }).bubbles is false
PASS new WheelEvent('eventType', { bubbles: true }).bubbles is true
PASS new WheelEvent('eventType', { cancelable: false }).cancelable is false
PASS new WheelEvent('eventType', { cancelable: true }).cancelable is true
PASS new WheelEvent('eventType', { view: window }).view is window
PASS new WheelEvent('eventType', { view: this }).view is this
PASS new WheelEvent('eventType', { view: testObject }).view threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { view: document }).view threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { view: undefined }).view is null
PASS new WheelEvent('eventType', { view: null }).view is null
PASS new WheelEvent('eventType', { view: false }).view threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { view: true }).view threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { view: '' }).view threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { view: 'chocolate' }).view threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { view: 12345 }).view threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { view: 18446744073709551615 }).view threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { view: NaN }).view threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { view: {valueOf: function () { return window; } } }).view threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { get view() { return 123; } }).view threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { get view() { throw 'WheelEvent Error'; } }) threw exception WheelEvent Error.
PASS new WheelEvent('eventType', { detail: 0 }).detail is 0
PASS new WheelEvent('eventType', { detail: -1 }).detail is -1
PASS new WheelEvent('eventType', { detail: 2147483647 }).detail is 2147483647
PASS new WheelEvent('eventType', { detail: -2147483648 }).detail is -2147483648
PASS new WheelEvent('eventType', { detail: 4294967295 }).detail is -1
PASS new WheelEvent('eventType', { detail: 9007199254740991 }).detail is -1
PASS new WheelEvent('eventType', { detail: 18446744073709551615 }).detail is 0
PASS new WheelEvent('eventType', { detail: 123.45 }).detail is 123
PASS new WheelEvent('eventType', { detail: NaN }).detail is 0
PASS new WheelEvent('eventType', { detail: undefined }).detail is 0
PASS new WheelEvent('eventType', { detail: null }).detail is 0
PASS new WheelEvent('eventType', { detail: '' }).detail is 0
PASS new WheelEvent('eventType', { detail: '12345' }).detail is 12345
PASS new WheelEvent('eventType', { detail: '12345a' }).detail is 0
PASS new WheelEvent('eventType', { detail: 'abc' }).detail is 0
PASS new WheelEvent('eventType', { detail: [] }).detail is 0
PASS new WheelEvent('eventType', { detail: [12345] }).detail is 12345
PASS new WheelEvent('eventType', { detail: [12345, 67890] }).detail is 0
PASS new WheelEvent('eventType', { detail: {} }).detail is 0
PASS new WheelEvent('eventType', { detail: {moemoe: 12345} }).detail is 0
PASS new WheelEvent('eventType', { detail: {valueOf: function () { return 12345; }} }).detail is 12345
PASS new WheelEvent('eventType', { screenX: 0 }).screenX is 0
PASS new WheelEvent('eventType', { screenX: -1 }).screenX is -1
PASS new WheelEvent('eventType', { screenX: 2147483647 }).screenX is 2147483647
PASS new WheelEvent('eventType', { screenX: -2147483648 }).screenX is -2147483648
PASS new WheelEvent('eventType', { screenX: 4294967295 }).screenX is -1
PASS new WheelEvent('eventType', { screenX: 9007199254740991 }).screenX is -1
PASS new WheelEvent('eventType', { screenX: 18446744073709551615 }).screenX is 0
PASS new WheelEvent('eventType', { screenX: 123.45 }).screenX is 123
PASS new WheelEvent('eventType', { screenX: NaN }).screenX is 0
PASS new WheelEvent('eventType', { screenX: undefined }).screenX is 0
PASS new WheelEvent('eventType', { screenX: null }).screenX is 0
PASS new WheelEvent('eventType', { screenX: '' }).screenX is 0
PASS new WheelEvent('eventType', { screenX: '12345' }).screenX is 12345
PASS new WheelEvent('eventType', { screenX: '12345a' }).screenX is 0
PASS new WheelEvent('eventType', { screenX: 'abc' }).screenX is 0
PASS new WheelEvent('eventType', { screenX: [] }).screenX is 0
PASS new WheelEvent('eventType', { screenX: [12345] }).screenX is 12345
PASS new WheelEvent('eventType', { screenX: [12345, 67890] }).screenX is 0
PASS new WheelEvent('eventType', { screenX: {} }).screenX is 0
PASS new WheelEvent('eventType', { screenX: {moemoe: 12345} }).screenX is 0
PASS new WheelEvent('eventType', { screenX: {valueOf: function () { return 12345; }} }).screenX is 12345
PASS new WheelEvent('eventType', { screenY: 0 }).screenY is 0
PASS new WheelEvent('eventType', { screenY: -1 }).screenY is -1
PASS new WheelEvent('eventType', { screenY: 2147483647 }).screenY is 2147483647
PASS new WheelEvent('eventType', { screenY: -2147483648 }).screenY is -2147483648
PASS new WheelEvent('eventType', { screenY: 4294967295 }).screenY is -1
PASS new WheelEvent('eventType', { screenY: 9007199254740991 }).screenY is -1
PASS new WheelEvent('eventType', { screenY: 18446744073709551615 }).screenY is 0
PASS new WheelEvent('eventType', { screenY: 123.45 }).screenY is 123
PASS new WheelEvent('eventType', { screenY: NaN }).screenY is 0
PASS new WheelEvent('eventType', { screenY: undefined }).screenY is 0
PASS new WheelEvent('eventType', { screenY: null }).screenY is 0
PASS new WheelEvent('eventType', { screenY: '' }).screenY is 0
PASS new WheelEvent('eventType', { screenY: '12345' }).screenY is 12345
PASS new WheelEvent('eventType', { screenY: '12345a' }).screenY is 0
PASS new WheelEvent('eventType', { screenY: 'abc' }).screenY is 0
PASS new WheelEvent('eventType', { screenY: [] }).screenY is 0
PASS new WheelEvent('eventType', { screenY: [12345] }).screenY is 12345
PASS new WheelEvent('eventType', { screenY: [12345, 67890] }).screenY is 0
PASS new WheelEvent('eventType', { screenY: {} }).screenY is 0
PASS new WheelEvent('eventType', { screenY: {moemoe: 12345} }).screenY is 0
PASS new WheelEvent('eventType', { screenY: {valueOf: function () { return 12345; }} }).screenY is 12345
PASS new WheelEvent('eventType', { clientX: 0 }).clientX is 0
PASS new WheelEvent('eventType', { clientX: -1 }).clientX is -1
PASS new WheelEvent('eventType', { clientX: 2147483647 }).clientX is 33554431
PASS new WheelEvent('eventType', { clientX: -2147483648 }).clientX is -33554432
PASS new WheelEvent('eventType', { clientX: 4294967295 }).clientX is -1
PASS new WheelEvent('eventType', { clientX: 9007199254740991 }).clientX is -1
PASS new WheelEvent('eventType', { clientX: 18446744073709551615 }).clientX is 0
PASS new WheelEvent('eventType', { clientX: 123.45 }).clientX is 123
PASS new WheelEvent('eventType', { clientX: NaN }).clientX is 0
PASS new WheelEvent('eventType', { clientX: undefined }).clientX is 0
PASS new WheelEvent('eventType', { clientX: null }).clientX is 0
PASS new WheelEvent('eventType', { clientX: '' }).clientX is 0
PASS new WheelEvent('eventType', { clientX: '12345' }).clientX is 12345
PASS new WheelEvent('eventType', { clientX: '12345a' }).clientX is 0
PASS new WheelEvent('eventType', { clientX: 'abc' }).clientX is 0
PASS new WheelEvent('eventType', { clientX: [] }).clientX is 0
PASS new WheelEvent('eventType', { clientX: [12345] }).clientX is 12345
PASS new WheelEvent('eventType', { clientX: [12345, 67890] }).clientX is 0
PASS new WheelEvent('eventType', { clientX: {} }).clientX is 0
PASS new WheelEvent('eventType', { clientX: {moemoe: 12345} }).clientX is 0
PASS new WheelEvent('eventType', { clientX: {valueOf: function () { return 12345; }} }).clientX is 12345
PASS new WheelEvent('eventType', { clientY: 0 }).clientY is 0
PASS new WheelEvent('eventType', { clientY: -1 }).clientY is -1
PASS new WheelEvent('eventType', { clientY: 2147483647 }).clientY is 33554431
PASS new WheelEvent('eventType', { clientY: -2147483648 }).clientY is -33554432
PASS new WheelEvent('eventType', { clientY: 4294967295 }).clientY is -1
PASS new WheelEvent('eventType', { clientY: 9007199254740991 }).clientY is -1
PASS new WheelEvent('eventType', { clientY: 18446744073709551615 }).clientY is 0
PASS new WheelEvent('eventType', { clientY: 123.45 }).clientY is 123
PASS new WheelEvent('eventType', { clientY: NaN }).clientY is 0
PASS new WheelEvent('eventType', { clientY: undefined }).clientY is 0
PASS new WheelEvent('eventType', { clientY: null }).clientY is 0
PASS new WheelEvent('eventType', { clientY: '' }).clientY is 0
PASS new WheelEvent('eventType', { clientY: '12345' }).clientY is 12345
PASS new WheelEvent('eventType', { clientY: '12345a' }).clientY is 0
PASS new WheelEvent('eventType', { clientY: 'abc' }).clientY is 0
PASS new WheelEvent('eventType', { clientY: [] }).clientY is 0
PASS new WheelEvent('eventType', { clientY: [12345] }).clientY is 12345
PASS new WheelEvent('eventType', { clientY: [12345, 67890] }).clientY is 0
PASS new WheelEvent('eventType', { clientY: {} }).clientY is 0
PASS new WheelEvent('eventType', { clientY: {moemoe: 12345} }).clientY is 0
PASS new WheelEvent('eventType', { clientY: {valueOf: function () { return 12345; }} }).clientY is 12345
PASS new WheelEvent('eventType', { wheelDeltaX: 0 }).wheelDeltaX is 0
PASS new WheelEvent('eventType', { wheelDeltaX: -1 }).wheelDeltaX is -1
PASS new WheelEvent('eventType', { wheelDeltaX: 2147483647 }).wheelDeltaX is 2147483647
PASS new WheelEvent('eventType', { wheelDeltaX: -2147483648 }).wheelDeltaX is -2147483648
PASS new WheelEvent('eventType', { wheelDeltaX: 4294967295 }).wheelDeltaX is -1
PASS new WheelEvent('eventType', { wheelDeltaX: 9007199254740991 }).wheelDeltaX is -1
PASS new WheelEvent('eventType', { wheelDeltaX: 18446744073709551615 }).wheelDeltaX is 0
PASS new WheelEvent('eventType', { wheelDeltaX: 123.45 }).wheelDeltaX is 123
PASS new WheelEvent('eventType', { wheelDeltaX: NaN }).wheelDeltaX is 0
PASS new WheelEvent('eventType', { wheelDeltaX: undefined }).wheelDeltaX is 0
PASS new WheelEvent('eventType', { wheelDeltaX: null }).wheelDeltaX is 0
PASS new WheelEvent('eventType', { wheelDeltaX: '' }).wheelDeltaX is 0
PASS new WheelEvent('eventType', { wheelDeltaX: '12345' }).wheelDeltaX is 12345
PASS new WheelEvent('eventType', { wheelDeltaX: '12345a' }).wheelDeltaX is 0
PASS new WheelEvent('eventType', { wheelDeltaX: 'abc' }).wheelDeltaX is 0
PASS new WheelEvent('eventType', { wheelDeltaX: [] }).wheelDeltaX is 0
PASS new WheelEvent('eventType', { wheelDeltaX: [12345] }).wheelDeltaX is 12345
PASS new WheelEvent('eventType', { wheelDeltaX: [12345, 67890] }).wheelDeltaX is 0
PASS new WheelEvent('eventType', { wheelDeltaX: {} }).wheelDeltaX is 0
PASS new WheelEvent('eventType', { wheelDeltaX: {moemoe: 12345} }).wheelDeltaX is 0
PASS new WheelEvent('eventType', { wheelDeltaX: {valueOf: function () { return 12345; }} }).wheelDeltaX is 12345
PASS new WheelEvent('eventType', { wheelDeltaY: 0 }).wheelDeltaY is 0
PASS new WheelEvent('eventType', { wheelDeltaY: -1 }).wheelDeltaY is -1
PASS new WheelEvent('eventType', { wheelDeltaY: 2147483647 }).wheelDeltaY is 2147483647
PASS new WheelEvent('eventType', { wheelDeltaY: -2147483648 }).wheelDeltaY is -2147483648
PASS new WheelEvent('eventType', { wheelDeltaY: 4294967295 }).wheelDeltaY is -1
PASS new WheelEvent('eventType', { wheelDeltaY: 9007199254740991 }).wheelDeltaY is -1
PASS new WheelEvent('eventType', { wheelDeltaY: 18446744073709551615 }).wheelDeltaY is 0
PASS new WheelEvent('eventType', { wheelDeltaY: 123.45 }).wheelDeltaY is 123
PASS new WheelEvent('eventType', { wheelDeltaY: NaN }).wheelDeltaY is 0
PASS new WheelEvent('eventType', { wheelDeltaY: undefined }).wheelDeltaY is 0
PASS new WheelEvent('eventType', { wheelDeltaY: null }).wheelDeltaY is 0
PASS new WheelEvent('eventType', { wheelDeltaY: '' }).wheelDeltaY is 0
PASS new WheelEvent('eventType', { wheelDeltaY: '12345' }).wheelDeltaY is 12345
PASS new WheelEvent('eventType', { wheelDeltaY: '12345a' }).wheelDeltaY is 0
PASS new WheelEvent('eventType', { wheelDeltaY: 'abc' }).wheelDeltaY is 0
PASS new WheelEvent('eventType', { wheelDeltaY: [] }).wheelDeltaY is 0
PASS new WheelEvent('eventType', { wheelDeltaY: [12345] }).wheelDeltaY is 12345
PASS new WheelEvent('eventType', { wheelDeltaY: [12345, 67890] }).wheelDeltaY is 0
PASS new WheelEvent('eventType', { wheelDeltaY: {} }).wheelDeltaY is 0
PASS new WheelEvent('eventType', { wheelDeltaY: {moemoe: 12345} }).wheelDeltaY is 0
PASS new WheelEvent('eventType', { wheelDeltaY: {valueOf: function () { return 12345; }} }).wheelDeltaY is 12345
PASS new WheelEvent('eventType', { deltaMode: 0 }).deltaMode is 0
PASS new WheelEvent('eventType', { deltaMode: 1 }).deltaMode is 1
PASS new WheelEvent('eventType', { deltaMode: 4294967294 }).deltaMode is 4294967294
PASS new WheelEvent('eventType', { deltaMode: 4294967295 }).deltaMode is 4294967295
PASS new WheelEvent('eventType', { deltaMode: WheelEvent.DOM_DELTA_PIXEL }).deltaMode is WheelEvent.DOM_DELTA_PIXEL
PASS new WheelEvent('eventType', { deltaMode: WheelEvent.DOM_DELTA_LINE }).deltaMode is WheelEvent.DOM_DELTA_LINE
PASS new WheelEvent('eventType', { deltaMode: WheelEvent.DOM_DELTA_PAGE }).deltaMode is WheelEvent.DOM_DELTA_PAGE
PASS new WheelEvent('eventType', { deltaMode: 9007199254740991 }).deltaMode is 4294967295
PASS new WheelEvent('eventType', { deltaMode: 18446744073709551615 }).deltaMode is 0
PASS new WheelEvent('eventType', { deltaMode: 12345678901234567890 }).deltaMode is 3944679424
PASS new WheelEvent('eventType', { deltaMode: -1 }).deltaMode is 4294967295
PASS new WheelEvent('eventType', { deltaMode: 123.45 }).deltaMode is 123
PASS new WheelEvent('eventType', { deltaMode: NaN }).deltaMode is 0
PASS new WheelEvent('eventType', { deltaMode: undefined }).deltaMode is 0
PASS new WheelEvent('eventType', { deltaMode: null }).deltaMode is 0
PASS new WheelEvent('eventType', { deltaMode: '' }).deltaMode is 0
PASS new WheelEvent('eventType', { deltaMode: '12345' }).deltaMode is 12345
PASS new WheelEvent('eventType', { deltaMode: '12345a' }).deltaMode is 0
PASS new WheelEvent('eventType', { deltaMode: 'abc' }).deltaMode is 0
PASS new WheelEvent('eventType', { deltaMode: [] }).deltaMode is 0
PASS new WheelEvent('eventType', { deltaMode: [12345] }).deltaMode is 12345
PASS new WheelEvent('eventType', { deltaMode: [12345, 67890] }).deltaMode is 0
PASS new WheelEvent('eventType', { deltaMode: {} }).deltaMode is 0
PASS new WheelEvent('eventType', { deltaMode: {moemoe: 12345} }).deltaMode is 0
PASS new WheelEvent('eventType', { deltaMode: {valueOf: function () { return 12345; }} }).deltaMode is 12345
PASS new WheelEvent('eventType', { ctrlKey: false }).ctrlKey is false
PASS new WheelEvent('eventType', { ctrlKey: true }).ctrlKey is true
PASS new WheelEvent('eventType', { altKey: false }).altKey is false
PASS new WheelEvent('eventType', { altKey: true }).altKey is true
PASS new WheelEvent('eventType', { shiftKey: false }).shiftKey is false
PASS new WheelEvent('eventType', { shiftKey: true }).shiftKey is true
PASS new WheelEvent('eventType', { metaKey: false }).metaKey is false
PASS new WheelEvent('eventType', { metaKey: true }).metaKey is true
PASS new WheelEvent('eventType', { button: 0 }).button is 0
PASS new WheelEvent('eventType', { button: 1 }).button is 1
PASS new WheelEvent('eventType', { button: 65534 }).button is 65534
PASS new WheelEvent('eventType', { button: 65535 }).button is 0
PASS new WheelEvent('eventType', { button: 9007199254740991 }).button is 0
PASS new WheelEvent('eventType', { button: -1 }).button is 0
PASS new WheelEvent('eventType', { button: 18446744073709551615 }).button is 0
PASS new WheelEvent('eventType', { button: 12345678901234567890 }).button is 2048
PASS new WheelEvent('eventType', { button: 123.45 }).button is 123
PASS new WheelEvent('eventType', { button: NaN }).button is 0
PASS new WheelEvent('eventType', { button: undefined }).button is 0
PASS new WheelEvent('eventType', { button: null }).button is 0
PASS new WheelEvent('eventType', { button: '' }).button is 0
PASS new WheelEvent('eventType', { button: '12345' }).button is 12345
PASS new WheelEvent('eventType', { button: '12345a' }).button is 0
PASS new WheelEvent('eventType', { button: 'abc' }).button is 0
PASS new WheelEvent('eventType', { button: [] }).button is 0
PASS new WheelEvent('eventType', { button: [12345] }).button is 12345
PASS new WheelEvent('eventType', { button: [12345, 67890] }).button is 0
PASS new WheelEvent('eventType', { button: {} }).button is 0
PASS new WheelEvent('eventType', { button: {moemoe: 12345} }).button is 0
PASS new WheelEvent('eventType', { button: {valueOf: function () { return 12345; }} }).button is 12345
PASS new WheelEvent('eventType', { relatedTarget: testDiv }).relatedTarget is testDiv
PASS new WheelEvent('eventType', { relatedTarget: document }).relatedTarget is document
PASS new WheelEvent('eventType', { relatedTarget: xhr }).relatedTarget is xhr
PASS new WheelEvent('eventType', { relatedTarget: testObject }) threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { relatedTarget: undefined }).relatedTarget is null
PASS new WheelEvent('eventType', { relatedTarget: null }).relatedTarget is null
PASS new WheelEvent('eventType', { relatedTarget: false }) threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { relatedTarget: true }) threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { relatedTarget: '' }) threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { relatedTarget: 'chocolate' }) threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { relatedTarget: 12345 }) threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { relatedTarget: 18446744073709551615 }) threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { relatedTarget: NaN }) threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { relatedTarget: {valueOf: function () { return testDiv; } } }) threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { get relatedTarget() { return 123; } }).relatedTarget threw exception TypeError: Type error.
PASS new WheelEvent('eventType', { get relatedTarget() { throw 'WheelEvent Error'; } }) threw exception WheelEvent Error.
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).bubbles is true
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).cancelable is true
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).view is window
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).detail is 111
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).screenX is 222
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).screenY is 333
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).clientX is 444
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).clientY is 555
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).ctrlKey is true
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).shiftKey is true
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).altKey is true
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).metaKey is true
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).button is 666
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).relatedTarget is testDiv
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).wheelDeltaX is 777
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).wheelDeltaY is 888
PASS new WheelEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, relatedTarget: testDiv, wheelDeltaX: 777, wheelDeltaY: 888, deltaMode: WheelEvent.DOM_DELTA_PAGE }).deltaMode is WheelEvent.DOM_DELTA_PAGE
PASS successfullyParsed is true

TEST COMPLETE
